Created paper
Made it in word, but have not git installed on this computer, so uploaded text only. Also, I do not know if you have Word, so this is better, only there are no linebreaks, which is kind of a bummer. Long story short I wrote the motivation and part of the 'how to do it and why'